Move autoloader to theme root
