--- inc/autoload.php
@@ -1,19 +1,0 @@
-<?php
-function WPChildThemeBoilerplateAutoload($class) {
-    $prefix = 'WPChildThemeBoilerplate\\';
-    $base_dir = __DIR__ . '/';
-
-    $len = strlen($prefix);
-    if (strncmp($prefix, $class, $len) !== 0) {
-        return;
-    }
-
-    $relative_class = substr($class, $len);
-    $file = $base_dir . str_replace('\\', '/', $relative_class) . '.php';
-
-    if (file_exists($file)) {
-        require $file;
-    }
-}
-
-spl_autoload_register('WPChildThemeBoilerplateAutoload');
